requirements_serverless.txt
# UniLife Backend - Serverless 部署依赖
# 精简版依赖，适合云函数部署

# ==================== 核心框架 ====================
fastapi==0.115.0
pydantic==2.9.2
pydantic-settings==2.6.6

# ==================== Serverless 适配 ====================
mangum==0.17.0

# ==================== 数据库 ====================
sqlalchemy==2.0.35
# PostgreSQL 异步驱动（云数据库必需）
asyncpg==0.30.0

# ==================== LLM 服务 ====================
openai==1.54.0
httpx==0.27.2

# ==================== 认证与安全 ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ==================== 时间处理 ====================
python-dateutil==2.9.0
pytz==2024.2

# ==================== HTTP 客户端 ====================
requests==2.32.5

# ==================== 注意事项 ====================
# 1. 移除了 uvicorn（云函数不需要）
# 2. 移除了 apscheduler（改用云函数定时触发器）
# 3. 移除了测试依赖（生产环境不需要）
# 4. 移除了 alembic（云函数不使用迁移工具）
